+++ spec/requests/as_content_owner/channel_spec.rb
@@ -131,6 +131,15 @@ describe Yt::Channel, :partner do
         end
       end
 
+      describe 'views can be grouped by related video' do
+        let(:range) { {since: 4.days.ago, until: 3.days.ago} }
+
+        specify 'with the :by option set to :related_video' do
+          views = channel.views range.merge by: :related_video
+          expect(views.keys).to all(be_instance_of Yt::Video)
+        end
+      end
+
       describe 'views can be grouped by video' do
         let(:range) { {since: 4.days.ago, until: 3.days.ago} }
 
@@ -598,6 +607,15 @@ describe Yt::Channel, :partner do
         specify 'with the :by option set to :embedded_player_location' do
           estimated_minutes_watched = channel.estimated_minutes_watched range.merge by: :embedded_player_location
           expect(estimated_minutes_watched).not_to be_empty
+        end
+      end
+
+      describe 'estimated minutes watched can be grouped by related video' do
+        let(:range) { {since: 4.days.ago, until: 3.days.ago} }
+
+        specify 'with the :by option set to :related_video' do
+          estimated_minutes_watched = channel.estimated_minutes_watched range.merge by: :related_video
+          expect(estimated_minutes_watched.keys).to all(be_instance_of Yt::Video)
         end
       end